requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apify_fingerprint_datapoints==0.11.0
browserforge==1.2.4
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cssselect==1.4.0
curl_cffi==0.14.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
msgspec==0.20.0
orjson==3.11.7
patchright==1.56.0
playwright==1.56.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
python-dotenv==1.2.2
scrapling==0.4.1
SQLAlchemy==2.0.48
starlette==0.52.1
tld==0.13.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
w3lib==2.4.0
win32_setctime==1.2.0
